+++ Interface/AddOns/LittleWigs/CHANGELOG.md
@@ -1,7 +1,7 @@
 # LittleWigs
 
-## [v7.0.7-3-g807ff0c](https://github.com/BigWigsMods/LittleWigs/tree/807ff0c4ff5dc28ada70b7f2ccb2df323e003a6d) (2016-09-07) [](#top)
-[Full Changelog](https://github.com/BigWigsMods/LittleWigs/compare/v7.0.7...807ff0c4ff5dc28ada70b7f2ccb2df323e003a6d)
+## [v7.0.8](https://github.com/BigWigsMods/LittleWigs/tree/v7.0.8) (2016-09-07) [](#top)
+[Full Changelog](https://github.com/BigWigsMods/LittleWigs/compare/v7.0.7...v7.0.8)
 
 -   BlackRookHold/KurtalosRavencrest: Fix options  
 -   VaultOfTheWardens/Tormentorum: Fix Lua error